backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
httpx>=0.27
pydantic>=2.11
python-multipart>=0.0.9
python-dotenv>=1.0.1
anthropic>=0.51.0
pytest>=8.2.0
